+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v12.0.6
 * Added "Class or Primary Resource" as an option to Resource Display icons.
+* Added new WoW 12.0.1 filters to the Buff/Debuff icon type.
 * Fix aura events not getting registered when tracking auras on many different units across all TMW icons.
 * Fix items sometimes showing with a ? texture.
 * Workaround a Blizzard bug where secret recharge durations for a max-charge ability are ignored.